+++ cypress/integration/options/verticalAxis.cy.js
@@ -6,7 +6,10 @@ import {
     clickDimensionModalUpdateButton,
 } from '../../elements/dimensionModal'
 import { goToStartPage } from '../../elements/startScreen'
-import { expectVisualizationToBeVisible } from '../../elements/chart'
+import {
+    expectVisualizationToBeVisible,
+    expectChartTitleToBeVisible,
+} from '../../elements/chart'
 import { TEST_DATA_ELEMENTS } from '../../utils/data'
 import { clickMenuBarOptionsButton } from '../../elements/menuBar'
 import {
@@ -22,10 +25,10 @@ import {
     setVerticalAxisTitle,
 } from '../../elements/optionsModal'
 import {
-    expectStoreConfigYAxisToHaveRangeMinValue,
-    expectStoreConfigYAxisToHaveRangeMaxValue,
-    expectStoreConfigYAxisToHaveTitleText,
-} from '../../utils/store'
+    expectWindowConfigYAxisToHaveRangeMinValue,
+    expectWindowConfigYAxisToHaveRangeMaxValue,
+    expectWindowConfigYAxisToHaveTitleText,
+} from '../../utils/window'
 import { generateRandomNumber, getRandomArrayItem } from '../../utils/random'
 
 const TEST_DATA_ELEMENT_NAME = getRandomArrayItem(TEST_DATA_ELEMENTS).name
@@ -54,9 +57,10 @@ describe('Options - Vertical axis', () => {
         })
         it('click the modal update button', () => {
             clickOptionsModalUpdateButton()
+            expectChartTitleToBeVisible()
         })
         it(`config has vertical axis title "${TEST_TITLE}"`, () => {
-            expectStoreConfigYAxisToHaveTitleText(TEST_TITLE)
+            expectWindowConfigYAxisToHaveTitleText(TEST_TITLE)
         })
     })
     describe('range', () => {
@@ -72,12 +76,13 @@ describe('Options - Vertical axis', () => {
         })
         it('click the modal update button', () => {
             clickOptionsModalUpdateButton()
+            expectChartTitleToBeVisible()
         })
         it(`config has range min value "${TEST_MIN_VALUE}"`, () => {
-            expectStoreConfigYAxisToHaveRangeMinValue(TEST_MIN_VALUE)
+            expectWindowConfigYAxisToHaveRangeMinValue(TEST_MIN_VALUE)
         })
         it(`config has range max value "${TEST_MAX_VALUE}"`, () => {
-            expectStoreConfigYAxisToHaveRangeMaxValue(TEST_MAX_VALUE)
+            expectWindowConfigYAxisToHaveRangeMaxValue(TEST_MAX_VALUE)
         })
     })
     // TODO: steps, decimals, labels