+++ packages/app/src/components/Layout/Layout.js
@@ -7,6 +7,7 @@ import {
     LAYOUT_TYPE_YEAR_OVER_YEAR,
     LAYOUT_TYPE_PIVOT_TABLE,
     getLayoutTypeByVisType,
+    LAYOUT_TYPE_SCATTER,
 } from '@dhis2/analytics'
 
 import DefaultLayout from './DefaultLayout/DefaultLayout'
@@ -14,12 +15,14 @@ import YearOverYearLayout from './YearOverYearLayout/YearOverYearLayout'
 import PieLayout from './PieLayout/PieLayout'
 import { sGetUiType } from '../../reducers/ui'
 import PivotTableLayout from './PivotTableLayout/PivotTableLayout'
+import ScatterLayout from './ScatterLayout/ScatterLayout'
 
 const componentMap = {
     [LAYOUT_TYPE_DEFAULT]: DefaultLayout,
     [LAYOUT_TYPE_PIE]: PieLayout,
     [LAYOUT_TYPE_YEAR_OVER_YEAR]: YearOverYearLayout,
     [LAYOUT_TYPE_PIVOT_TABLE]: PivotTableLayout,
+    [LAYOUT_TYPE_SCATTER]: ScatterLayout,
 }
 
 const Layout = ({ visType }) => {