+++ packages/app/src/modules/current.js
@@ -13,11 +13,16 @@ import {
     getPredefinedDimensionProp,
     DIMENSION_PROP_NO_ITEMS,
     getAdaptedUiLayoutByType,
+    VIS_TYPE_SCATTER,
 } from '@dhis2/analytics'
 
 import options from './options'
 import {} from './layout'
 import { BASE_FIELD_TYPE, BASE_FIELD_YEARLY_SERIES } from './fields/baseFields'
+import {
+    ITEM_ATTRIBUTE_HORIZONTAL,
+    ITEM_ATTRIBUTE_VERTICAL,
+} from '../modules/ui'
 
 const hasItems = (object, id) => Array.isArray(object[id]) && object[id].length
 
@@ -125,6 +130,39 @@ export const getSingleValueCurrentFromUi = (state, action) => {
         ...state,
         [BASE_FIELD_TYPE]: ui.type,
         ...singleValueAxesFromUi,
+        ...getOptionsFromUi(ui),
+    }
+}
+
+export const getScatterCurrentFromUi = (state, action) => {
+    const ui = {
+        ...action.value,
+        layout: {
+            ...getAdaptedUiLayoutByType(action.value.layout, VIS_TYPE_SCATTER),
+        },
+    }
+
+    const axesFromUi = getAxesFromUi(ui)
+
+    // only save first vertical and first horizontal dx items
+    const verticalItem =
+        ui.itemAttributes.find(
+            item => item.attribute === ITEM_ATTRIBUTE_VERTICAL
+        ) || {}
+    const horizontalItem =
+        ui.itemAttributes.find(
+            item => item.attribute === ITEM_ATTRIBUTE_HORIZONTAL
+        ) || {}
+    const scatterAxesFromUi = layoutReplaceDimension(
+        axesFromUi,
+        DIMENSION_ID_DATA,
+        [verticalItem, horizontalItem].map(item => ({ id: item.id }))
+    )
+
+    return {
+        ...state,
+        [BASE_FIELD_TYPE]: ui.type,
+        ...scatterAxesFromUi,
         ...getOptionsFromUi(ui),
     }
 }